+++ packages/mirror-media-next/components/login/form-wrapper.js
@@ -4,9 +4,14 @@ import Wrapper from '../shared/form-wrapper'
 
 const FormWrapper = styled(Wrapper)`
   ${({ theme }) => theme.breakpoint.md} {
-    width: 468px;
+    width: 423px;
     margin-top: 48px;
     margin-bottom: 48px;
+  }
+
+  ${({ theme }) => theme.breakpoint.xxl} {
+    margin-top: 60px;
+    margin-bottom: 60px;
   }
 `
