+++ packages/mirror-media-next/components/login/main-form.js
@@ -2,7 +2,7 @@ import styled, { css } from 'styled-components'
 import { useAppSelector } from '../../hooks/useRedux'
 import { loginFormMode, FormMode } from '../../slice/login-slice'
 import NextImage from 'next/image'
-import FormWrapper from '../shared/form-wrapper'
+import FormWrapper from './form-wrapper'
 import MainFormStart from './main-form-start'
 import MainFormRegistration from './main-form-registration'
 import MainFormLoginWithPassword from './main-form-login-with-password'
