+++ packages/mirror-media-next/components/shared/form-wrapper.js
@@ -10,18 +10,10 @@ const FormWrapper = styled.div`
 
   ${({ theme }) => theme.breakpoint.md} {
     padding: 40px 40px;
-    margin-top: 48px;
-    margin-bottom: 48px;
-    width: 423px;
     border-radius: 24px;
     border: 1px solid rgba(0, 0, 0, 0.1);
     box-shadow: 0px 2px 12px 0px rgba(0, 0, 0, 0.08),
       0px 4px 28px 0px rgba(0, 0, 0, 0.06);
-  }
-
-  ${({ theme }) => theme.breakpoint.xxl} {
-    margin-top: 60px;
-    margin-bottom: 60px;
   }
 `
