+++ packages/mirror-media-next/pages/recover-password.js
@@ -4,7 +4,7 @@ import { setPageCache } from '../utils/cache-setting'
 import { GCP_PROJECT_ID, ACTION_CODE_SETTING } from '../config/index.mjs'
 import { InputState } from '../constants/form'
 import LayoutFull from '../components/shared/layout-full'
-import FormWrapper from '../components/shared/form-wrapper'
+import FormWrapper from '../components/login/form-wrapper'
 import GenericTextInput from '../components/shared/inputs/generic-text-input'
 import PrimaryButton from '../components/shared/buttons/primary-button'
 import TextButton from '../components/login/text-button'
